requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
email-validator==2.1.0.post1
celery==5.3.6
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
openai==1.3.7
anthropic==0.7.4
python-socketio==5.10.0 